fix: Error on saving money field
When saving a DataObject containing a money field, it errors out on saving, because the number typed in into the field is parsed as a string to the PHP function. That is easily mitigated by parsing the string into a float, because the function already checks if the string is numeric.